+++ client/hawc_client/assessment.py
@@ -2,7 +2,7 @@ from typing import Dict, List
 
 import pandas as pd
 
-from .base_client import BaseClient
+from .client import BaseClient
 
 
 class AssessmentClient(BaseClient):
